+++ plugin/export_scripts_template/autoloads/sign_in_client.gd
@@ -13,6 +13,11 @@ extends Node
 ## [param is_authenticated]: Indicates if the user is authenticated or not.
 signal user_authenticated(is_authenticated: bool)
 
+## Signal emitted after calling the [method request_server_side_access] method.[br]
+## [br]
+## [param token]: The OAuth 2.0 authorization code as a string.
+signal server_side_access_requested(token:String)
+
 ## This signal was emitted after calling the [method sign_in] method, but it's not emitted
 ## anymore. Instead, the [signal user_authenticated] signal is emitted now.[br]
 ## [br]
@@ -27,6 +32,9 @@ func _connect_signals() -> void:
 	if GodotPlayGameServices.android_plugin:
 		GodotPlayGameServices.android_plugin.userAuthenticated.connect(func(is_authenticated: bool):
 			user_authenticated.emit(is_authenticated)
+		)
+		GodotPlayGameServices.android_plugin.serverSideAccessRequested.connect(func(token: String):
+			server_side_access_requested.emit(token)
 		)
 
 ## Use this method to check if the user is already authenticated. If the user is authenticated,
@@ -43,3 +51,16 @@ func is_authenticated() -> void:
 func sign_in() -> void:
 	if GodotPlayGameServices.android_plugin:
 		GodotPlayGameServices.android_plugin.signIn()
+
+## Requests server-side access to Play Games Services for the currently signed-in player.
+## When requested, an authorization code is returned that can be used by your server to exchange
+## for an access token that can be used by your server to access the Play Games Services web APIs.[br]
+## [br]
+## The method emits the [signal server_side_access_requested] signal.[br]
+## [br]
+## [param server_client_id]: The client ID of the server that will perform the authorization code flow exchange.[br]
+## [param force_refresh_token]: If true, when the returned authorization code is exchanged, a refresh
+## token will be included in addition to an access token.
+func request_server_side_access(server_client_id: String, force_refresh_token: bool) -> void:
+	if GodotPlayGameServices.android_plugin:
+		GodotPlayGameServices.android_plugin.requestServerSideAccess(server_client_id, force_refresh_token)